requirements.txt
python-dotenv
imap-tools
playwright
requests
